fix delete file

+++ util/file.js
@@ -1,9 +1,11 @@
 const fs = require("fs");
 
 const deletFile = (path) => {
-  fs.unlink(path, (err) => {
-    if (err) throw err;
-  });
+  if (fs.existsSync(path)) {
+    fs.unlink(path, (err) => {
+      if (err) throw err;
+    });
+  }
 };
 
 module.exports = deletFile;
